+++ R/scale_process.R
@@ -122,34 +122,37 @@ par_grid <-
     funname <- as.character(substitute(fun_dist))
     # is the function extract_at?
     is_extract_at <- any(endsWith(funname, "extract_at"))
+    message(is_extract_at)
     funname <- funname[length(funname)]
     pkgname <- try(.check_package(funname), silent = TRUE)
 
     # parallel worker will take terra class objects
     # if chopin function is used
-    class_vec <-
-      if (pkgname == "chopin") {
-        if (is_extract_at) {
-          "sf"
-        } else {
-          "terra"
-        }
-      } else {
-        pkgname
-      }
+    class_vec <- "sf"
+      # if (pkgname == "chopin") {
+        # if (is_extract_at) {
+        #   "sf"
+        # } else {
+        #   "terra"
+        # }
+      # } else {
+      #   pkgname
+      # }
 
     # clean additional arguments
     args_input <- list(...)
-    if (funname == "chopin" && is.null(args_input$.standalone)) {
+    if (pkgname == "chopin" && is.null(args_input$.standalone)) {
       args_input$.standalone <- FALSE
     }
 
     # Track spatraster file path
     args_input$x <- .check_par_spatraster(args_input$x)
     args_input$y <- .check_par_spatraster(args_input$y)
+
     # get hints from the inputs on data model
     peek_x <- try(.check_character(args_input$x), silent = TRUE)
     peek_y <- try(.check_character(args_input$y), silent = TRUE)
+
     if (inherits(peek_x, "try-error")) {
       crs_x <- terra::crs(args_input$x)
     } else {
@@ -392,15 +395,15 @@ par_hierarchy <-
     # parallel worker will take terra class objects
     # if chopin function is used
     class_vec <-
-      if (pkgname == "chopin") {
+      # if (pkgname == "chopin") {
         if (is_extract_at) {
           "sf"
         } else {
           "terra"
         }
-      } else {
-        pkgname
-      }
+      # } else {
+      #   pkgname
+      # }
 
     # Track spatraster file path
     args_input$x <- .check_par_spatraster(args_input$x)
@@ -702,15 +705,15 @@ par_multirasters <-
     # parallel worker will take terra class objects
     # if chopin function is used
     class_vec <-
-      if (pkgname == "chopin") {
+      # if (pkgname == "chopin") {
         if (is_extract_at) {
           "sf"
         } else {
           "terra"
         }
-      } else {
-        pkgname
-      }
+      # } else {
+      #   pkgname
+      # }
 
     # Unlike other par_* functions, raster paths are not
     # tracked by the function since the raster file paths