+++ NEWS.md
@@ -4,6 +4,7 @@
 - CRAN ready: unexported function examples are moved to /tools on GitHub
 - SoftwareX paper is published; DOI and inst/CITATION is added
 - Supports H3 and DGGRID gridding functions from 0.9.5
+- Generic function using all-`sf` inputs in `x` and `y` is supported in `par_*()` functions
 
 # 0.8
 - Bumped version from 0.7.8 to 0.8.0: improving package coverage